refactor(actuators): consolidate errors
Consolidate errors related to actuators e.g. experiments in catalogs/registry, actuators in registry, measurement error
